Docker_disease-detection-service/requirements.txt
flask
flask-cors
numpy
tensorflow
Pillow
opencv-python-headless

